+++ R/graphs.R
@@ -331,8 +331,6 @@ rebound_graphs_helper <- function(.path_data,
 #'                    See examples.
 #' @param line_var The name of variable to be used to discriminate lines on the graph. 
 #'                 Default is `y_names_col`.
-#' @param include_x_axis A boolean that tells whether to include a horizontal line at y = 0.
-#'                       Default is `FALSE`.
 #' @param graph_params A list of parameters to control graph appearance. 
 #'                     See `ReboundTools::sens_graph_params`.
 #' @param y_names_col,y_vals_col See `ReboundTools::graph_df_colnames`.
@@ -437,7 +435,6 @@ sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_da
                                x_var,
                                y_var,
                                line_var = y_names_col,
-                               include_x_axis = FALSE,
                                y_names_col = ReboundTools::graph_df_colnames$y_names_col,
                                y_vals_col = ReboundTools::graph_df_colnames$y_vals_col,
                                graph_params = ReboundTools::sens_graph_params,
@@ -463,14 +460,14 @@ sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_da
 
   # Create the graph and return it
   g <- ggplot2::ggplot()
-  if (include_x_axis) {
+  if (graph_params$include_x_axis) {
     g <- g +
       ggplot2::geom_hline(yintercept = 0, size = 0.2)
   }
   g + 
     ggplot2::geom_point(data = orig_data,
-                        mapping = ggplot2::aes_string(x = x_var, y = y_vals_col),
-                        colour = graph_params$orig_point_colour,
+                        # Pick up the dot colour from the line colour.
+                        mapping = ggplot2::aes_string(x = x_var, y = y_vals_col, colour = line_var),
                         size = graph_params$orig_point_size,
                         shape = graph_params$orig_point_shape,
                         stroke = graph_params$orig_point_stroke,
@@ -518,8 +515,6 @@ sensitivity_graphs <- function(.parametric_data = parametric_analysis(rebound_da
 #' @param point_type_colname,sweep_points,orig_points See `ReboundTools::parametric_analysis_point_types`.
 #' @param line_var The name of variable to be used to discriminate lines on the graph. 
 #'                 Default is `y_names_col`.
-#' @param include_x_axis A boolean that tells whether to include a horizontal line at y = 0.
-#'                       Default is `TRUE`.
 #' @param y_names_col,y_vals_col See `ReboundTools::graph_df_colnames`.
 #' @param graph_params A list of parameters to control graph appearance. 
 #'                     See `ReboundTools::sens_graph_params`.
@@ -542,7 +537,6 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                 x_var,
                                 Re_terms = unlist(ReboundTools::rebound_terms),
                                 line_var = y_names_col,
-                                include_x_axis = TRUE,
                                 y_names_col = ReboundTools::graph_df_colnames$y_names_col,
                                 y_vals_col = ReboundTools::graph_df_colnames$y_vals_col,
                                 graph_params = ReboundTools::sens_graph_params,
@@ -550,12 +544,20 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                 sweep_points = ReboundTools::parametric_analysis_point_types$sweep,
                                 orig_points = ReboundTools::parametric_analysis_point_types$orig) {
   Re_terms <- match.arg(Re_terms, several.ok = TRUE)
+  
+  if (graph_params$use_latex_legend) {
+    legend_labs <- ReboundTools::latex_rebound_terms[Re_terms] %>% 
+      latex2exp::TeX() %>% 
+      unlist()
+  } else {
+    legend_labs <- Re_terms %>% unlist()
+  }
+  
 
   sensitivity_graphs(.parametric_data = .parametric_data, 
                      x_var = x_var, 
                      y_var = Re_terms, 
                      line_var = line_var, 
-                     include_x_axis = include_x_axis,
                      y_vals_col = y_vals_col,
                      y_names_col = y_names_col,
                      graph_params = graph_params, 
@@ -577,6 +579,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                             Re_dir = graph_params$dir_colour,
                                             Re_indir = graph_params$indir_colour,
                                             Re_tot = graph_params$tot_colour), 
+                                 labels = legend_labs,
                                  breaks = Re_terms) +
     ggplot2::scale_size_manual(values = c(Re_dempl = graph_params$dempl_size,
                                           Re_emb = graph_params$emb_size,
@@ -593,6 +596,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                           Re_dir = graph_params$dir_size,
                                           Re_indir = graph_params$indir_size,
                                           Re_tot = graph_params$tot_size),
+                               labels = legend_labs,
                                breaks = Re_terms) +
     ggplot2::scale_linetype_manual(values = c(Re_dempl = graph_params$dempl_linetype, 
                                               Re_emb = graph_params$emb_linetype,
@@ -609,6 +613,7 @@ rebound_terms_graph <- function(.parametric_data = parametric_analysis(rebound_d
                                               Re_dir = graph_params$dir_linetype,
                                               Re_indir = graph_params$indir_linetype,
                                               Re_tot = graph_params$tot_linetype), 
+                                   labels = legend_labs,
                                    breaks = Re_terms)
 }